+++ src/lib/uuid.ts
@@ -1,5 +1,5 @@
 /* eslint-disable no-mixed-operators */
-export default function uuid(): string {
+export function uuid(): string {
   return "xxxxxxxx-xxxx-4xxx-yxxx-xxxxxxxxxxxx".replace(/[xy]/g, function (c) {
     const r = (Math.random() * 16) | 0;
     const v = c === "x" ? r : (r & 0x3) | 0x8;